+++ src/commands/xp-role/list/text.ts
@@ -1,9 +1,7 @@
-import Config from "adapters/config"
-import { APIError, GuildIdNotFoundError } from "errors"
-import { list } from "../processor"
 import { Command } from "types/common"
 import { PREFIX, XP_ROLE_GITBOOK } from "utils/constants"
 import { composeEmbedMessage } from "ui/discord/embed"
+import { process } from "./processor"
 
 const command: Command = {
   id: "xr_list",
@@ -11,31 +9,7 @@ const command: Command = {
   brief: "List all the xp role setup",
   category: "Config",
   onlyAdministrator: true,
-  run: async function (msg) {
-    if (!msg.guildId || !msg.guild) {
-      throw new GuildIdNotFoundError({ message: msg })
-    }
-    const res = await Config.getConfigXPRoleList(msg.guildId)
-    if (!res.ok) {
-      throw new APIError({
-        message: msg,
-        curl: res.curl,
-        description: res.log,
-      })
-    }
-
-    const { title, description } = list(res)
-    return {
-      messageOptions: {
-        embeds: [
-          composeEmbedMessage(msg, {
-            author: [title],
-            description,
-          }),
-        ],
-      },
-    }
-  },
+  run: process,
   getHelpMessage: async (msg) => ({
     embeds: [
       composeEmbedMessage(msg, {
